refactor: tornando o filtro de exceção global flexivel
