+++ src/controller/controllers/auth.controller.ts
@@ -3,7 +3,7 @@ import { AuthService } from '@/model/services/auth.service'
 import { AuthLoginDto } from '@/model/entities/dto/auth-login.dto'
 import { HashPasswordPipe } from '@/model/common/pipes/hash-password.pipe'
 import { ApiBody, ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger'
-import { AuthRegisterDto } from '@/model/entities/dto//auth-register.dto'
+import { AuthRegisterDto } from '@/model/entities/dto/auth-register.dto'
 
 @Controller('auth')
 @ApiTags('auth')
