+++ src/model/common/filters/global-exception.filter.ts
@@ -1,0 +1,37 @@
+import {
+  ExceptionFilter,
+  Catch,
+  ArgumentsHost,
+  HttpException,
+  HttpStatus,
+} from '@nestjs/common'
+import { HttpAdapterHost } from '@nestjs/core'
+
+@Catch()
+export class GlobalExceptionFilter implements ExceptionFilter {
+  constructor(private adapterHost: HttpAdapterHost) {}
+  catch(exception: unknown, host: ArgumentsHost) {
+    const { httpAdapter } = this.adapterHost
+
+    const ctx = host.switchToHttp()
+    const response = ctx.getResponse()
+    const request = ctx.getResponse()
+
+    const { status, body } =
+      exception instanceof HttpException
+        ? {
+            status: exception.getStatus(),
+            body: exception.getResponse(),
+          }
+        : {
+            status: HttpStatus.INTERNAL_SERVER_ERROR,
+            body: {
+              statusCode: HttpStatus.INTERNAL_SERVER_ERROR,
+              timestamp: new Date().toISOString(),
+              path: httpAdapter.getRequestUrl(request),
+            },
+          }
+
+    httpAdapter.reply(response, body, status)
+  }
+}
